Add a cache for HW registers
HW registers (status, mode, trapsts, pc, ...) are often accessed by
the debugger API library to inspect the state of the wavefront.

To avoid making multiple successive syscalls to read the content of
the context save area in the target's memory, we read all the hwregs
in one chunk, and keep them cached. The cache is valid as long as
the wave is stopped.

Added a minimal std::optional implementation (only available in C++17)
to conditionally suspend the queue if reading a non-cached register.

Change-Id: I8fffe31f29a008b9a3356e7a1fe74db66bcbdfe1


[ROCm/rocdbgapi commit: 6248242eb49d31cee3c4adbe6747c8df22b802e2]