Fix single-step for s_trap instructions
The spurious event detection prevents s_trap instructions from
being single-stepped. We already handle the case of branch to self,
so we just need to extend it to handle s_trap as well.

Also, we "park" waves at breakpoint or engpgm instruction, but a
few cases were missed. This commit addresses that.

Lastly, when we detach, we should not resume waves halted because of
exceptions.  Keep those halted and if the debugger is re-attached
it will report the exceptions again.

Change-Id: Iada8b7deb3e61e3abc74904bc0480caa998bc12f


[ROCm/rocdbgapi commit: d2ee25b1d0f741d3cbd8d02dd524a7834850483b]