Add a no_agent_driver_t class
The no_agent_driver_t instance is used if the kfd driver is not
available.  This enables the debugger library, but no agents and
no queues are ever created.

Moved agent_t discovery to os_driver_t::agent_snapshot so that it
is not part of the OS driver layer.

Change-Id: Iaf84259fd89580ad9bab2a31e6b2df68a6119a82


[ROCm/rocdbgapi commit: 38d5bea28b18a81a08d6097dcf6b45ceb4809b2a]