Fix an issue with invalid queues
Queues are not necessarily deleted as soon as they become invalid, so
it is possible for the queue map to contain multiple entries with the
same os_queue_id, all but one being invalid.  Because of this,
handle_object_set_t::find_if should exclude invalid objects from the
search.

Moved the waves and dispatches destruction when a queue is invalidated
to the queue_impl_t destructor. Only queues creating waves should have
to do the cleanup.

Added some log-info messages for queue creation/invalidation/deletion.

Change-Id: Ibc99d9fca3f8984afd9198a79a412d0007e5f17e


[ROCm/rocdbgapi commit: 9bdeb43193a6a07d8d6620bd72ecdffc07f82452]