Fix error "trap raised should only ..."
It is possible for a breakpoint instruction to be removed before
inspecting the s_trap instruction's immediate.  The trap handler
should return the stop reason to avoid any data races, but until
then, assume the trap_id is 7 (breakpoint instruction) if the
instruction at the stopped pc is not an s_trap.

Change-Id: I67671bbb98fca4c4c0a1e36d72c9f7e175499f43


[ROCm/rocdbgapi commit: e4d3ba219b04590c73f2d9ccb986e04f1c7a8202]