Don't return an event if r_state != RT_CONSISTENT
When the code object list is being updated by the ROCR, the
r_debug::r_state is set to r_debug::RT_ADD. The r_brk function might
still be called, triggering a breakpoint, but there's no need to report
the event to the debugger since we can't update the code object list.

Remove limiting the size of reading a code object URI to avoid future
issues where the maximum size may increase.

Change-Id: Iec2b1e0bc88839fb55039fdc932f03f86a83cf84


[ROCm/rocdbgapi commit: bd36b05d7b9e4bcc53b978dd1ecaddf6dd01edbf]