Displaced stepping fixes
amd_dbgapi_displaced_stepping_start should return a memory error instead
of a fatal error if it fails to read from the wave's current pc.

Return an error if the displaced instruction is not recognized by the
disassembler.  It may be non-sequencial and we can't tell if the jump is
relative or absolute.

Change-Id: Ic51fe7e2ee8d2c49b487691eaf506c819aac4de1


[ROCm/rocdbgapi commit: d8342e07beb11c1d873b6a545ac318b5608d6534]