Update monotonic_counter and reset global IDs
Removed the prefix and postfix increment operators and replaced them
with a function operator to access the underlying counter.  The counter
is guaranted to be incremented each time it is accessed.

Reset the process' handle_object IDs when the library instance is
finalized.  This means that the next instance can reuse the same
global IDs, making ID assignment more deterministic.

Change-Id: I27918c060347fddcd2ff6fe46eb111bcb5c8f993


[ROCm/rocdbgapi commit: 0ab4395949f6990f14d37f685ce8af36039dc52f]