Use the new process-wide kfd_driver_t ioctls
Moved the per agent polling file descriptor to a single file descriptor,
global to the process.  KFD writes to the process file descriptor when
a queue event is pending.

Moved debug enable/disable to the process.  All agents are enabled or
disabled by the KFD at once.

Updated kfd_driver_t to use the new KFD ioctls using the PID instead of
the gpu_id.

Moved the event thread to the os_driver_t class.

Change-Id: Ie3d436255369e1417442e67dd25c5b34398ca031


[ROCm/rocdbgapi commit: ed75d67ceeb32d4ad8341dd325202cd3e9e58025]