Fix displaced stepping completion on detach
When the library is detached from the process, it needs to cancel all
outstanding displaced step operations.  To successfully complete a
displaced stepping, the wave must be stopped and the queue must be
suspended.

Suspending the queue in the wave_t destructor is problematic (suspend
may call queue snapshot, queue snapshot may destruct the queue and queue
impl, and the wave may be left with a dangling queue reference). So,
instead, complete the displaced stepping in process_t::detach where the
state of the queue is known to be suspended.

Change-Id: I705dbdac611e88ee36bae026e2b0fd38f02582a5


[ROCm/rocdbgapi commit: aa13e1c173856fd0323310f184e9ca3a71c208dc]