Fix amd_dbgapi_wave_stop
Add state to wave_t to record whether there is an outstanding stop
request.

Return an error if the client tries to stop a wave that is already
stop or has an outstanding stop request.

Generate a command terminated event if the wave had not completed
a single-step when requested to stop.

Change-Id: I3323d9483df509cbcb1f0ac31709a27974931c13


[ROCm/rocdbgapi commit: d662cd5e53b154f7c598dccb5c1de080062739fa]