Add a state to event_t
Track the progress of an event_t from the time it is created to the
time it is destroyed.  The state of an event_t can only progress from
created->queued->reported->processed.  Once the state reaches
processed, the event_t may be destroyed.

An event is marked as reported when it is returned from
amd_dbgapi_process_next_pending_event.  It is marked as processed when
the client calls amd_dbgapi_event_processed.

The event state enables the correct reporting of the client visible
wave state when the wave is stopped but the client has not retrieved
the event and so does not yet know about it.

Change-Id: Iad586c046e7e7318d1ae0b84bc09e5ba35b724e6


[ROCm/rocdbgapi commit: eb2841aca47097254a17e41a86f3d2b4b8439d01]