Make stopped and halted distinct states
A wave can be stopped as the result of entering the trap handler
due to a trap or exception, or because of a user request.

A wave can be halted by the trap handler to implement the stopped
state, or after executing a s_sethalt 1 instruction.

A wave can be both running (not stopped), and halted (not executing
instructions).

Change-Id: I8e94a36dfd5f84e2e4879bea1b202d3eba733ab2


[ROCm/rocdbgapi commit: 46de96402a9ff78e3ef1f540964f9c0509fdc44f]