Release instruction buffers if the wave is running
queue_t::update_waves creates new wave_t instances as they are
discovered in the queue's context save area before old stale wave
instances are deleted.

Because of this, it is possible to have more wave_t instances than
the hardware can launch waves, and if all instances hold an instruction
buffer, the queue could run out of buffers to park new waves.

The solution is to release a wave's instruction buffer before it is
resumed in normal mode.

Change-Id: I01df691ea612146c2586b9bb857c2fac83314aed


[ROCm/rocdbgapi commit: 3eed5680ec28c19164c1eeafadf490c045abf075]