Remove unused wave_t instruction buffer
Since parking a wave now uses a shared instruction buffer owned by the
queue (same for the endpgm instruction buffer), the wave no longer needs
an instruction buffer.

When a displaced_stepping_t is created, a new instruction buffer is
allocated and used until the displaced_stepping_t is destructed.

Change-Id: Id41c18352c720e6e6de2ee074dc2c1282e648f19


[ROCm/rocdbgapi commit: 27f8e11af06c3722af185c667c07fe2f95350aaa]