Add support for cached global memory accesses
Add a memory_cache_t to cache frequent access to global memory. The
supported cache policies are uncached, write-through and write-back.

Use a write-back memory_cache_t for caching a wave's hwregs.  The
hwregs cache is reset each time the queue is suspended, and flushed
when the queue is resumed.

Change-Id: Ibbd7594c4da7771cf0c4466412bed49ea1b9bfe2


[ROCm/rocdbgapi commit: bd487fa63d9ad49fe4731d299b02b2ec5487bf90]