Add utils::doubly_linked_list_t
A doubly_linked_entry_t<T> can be embedded in an object of type T,
avoiding the need to allocate a node to hold a reference to  the
linked element.

A doubly_linked_list_t<T> contains a doubly_linked_entry_t<T> head,
from which entries can be iterated forward or backard.

Change-Id: I2362db3a5025e9eb1291246128a6e249b075f9b1


[ROCm/rocdbgapi commit: d8060355304e1c07c84a7db227cd08ec5e5df1a0]