Don't report code object events on attach/detach
The code objects updated event needs to be reported only when it is
guaranted that a CPU thread is stopped, for example, when the ROCR hits
the rbrk breakpoint after loading a new code object.  Without a stopped
CPU thread, the debugger cannot read/write memory using ptrace.

After attaching, it is the debugger's responsibility to get the
initial list of loaded code objects.  The code object updated event
will only be reported if that list changes.

Change-Id: I6eca424b763593626c4f1c0f6d948e7787e2dcb9


[ROCm/rocdbgapi commit: 3055fc00b0689feace96deba7051211f1df07d1b]