Advance the PC after trap exceptions
To make it easier to handle user inserted trap instructions (llvm.trap,
or llvm.debugtrap), advance the PC after stopping because of an
exception raised by a trap instruction. The user can then simply resume
the wave's execution clear of the trap instruction.

For breakpoint traps, it is the debugger's responsibility to rewind the
PC to point at the instruction, not past the instruction, using the
breakpoint pc adjusment value returned by the
AMD_DBGAPI_ARCHITECTURE_INFO_BREAKPOINT_INSTRUCTION_PC_ADJUST
architecture query.

Change-Id: Ia3f7a90aa27e3fb5b19110a53d56521223881532


[ROCm/rocdbgapi commit: 4ee0f377a2f518c9e67524a83059f2b57e846184]