Implement resuming a wave with an exception
When resuming a wave, an exception can be raised which causes the
runtime handler to be invoked.  When an exception is raised, the
wave is halted (status.halt=1) just like it would if the debugger
was not attached.

The runtime, or a user registered handler,  may put the queues in
the error state and terminate the application.

Change-Id: I125ed9d9cd485c7b9c7053c8c1f49231bc9ad0d6


[ROCm/rocdbgapi commit: 6b90315ee66ca08ac187d2595ecc6d76aaaa2977]