Add support for device memory violations
Subscribe to the device_memory_violation events.

A device memory violation could be caused by a wave memory violation
(in which case the dbgapi will also receive a queue_memory_violation
event), the command procesor, or the DMA engine.

If a device memory violation event is received, the library looks for
a wave, on the device, with a memory violation exception (trapsts
.excp.mem_viol, trapsts.xnack_error) and if none are found, the
exception is immediately forwarded to the runtime.

If a wave with a memory violation exception is resumed and a device
memory violation exception is recorded, the runtime is sent a device
memory violation exception instead of a queue memory violation.

When resuming a wave, all exceptions are cleared.

Change-Id: I2b03bb415ebeb9d53271b51bf812db62c94697d4


[ROCm/rocdbgapi commit: 8784a1b2bf2dc02f0fb291eef439c715e7c78825]