Update os_driver::send_exceptions
It is possible to send exceptions that span the process, agent and
queue groups, so the ioctl is now taking both a queue_id and an agent_id
(process does not require an id).

If any exception in the queue mask is set, then the os_queue_id must be
set, and if any exception in agent mask is set then the os_agent_id must
be set.

Remove the FIXME and temporary workaround to limit the source to only
one exception group.

Change-Id: Ic4a38747181742892031aa1f5132544df330381e


[ROCm/rocdbgapi commit: 09a3668309543e0dbea764bfdb79ccd5741d6527]