Improve instruction simulation
Require a wave to be in single-step mode before starting instruction
simulation.  After simulating the instruction, also simulate executing
the trap handler (save the status and pc, and park and halt the wave).

Remove instruction simulation in the spurious event detection, and
instead simulate non-sequential instructions when the wave is resumed
in single-step mode.

Change-Id: Ib58ae6ba1464e3f9a7ad59fc956756435451241b


[ROCm/rocdbgapi commit: d8f500f1799b10ae1e364bd000ac8d9a45a6b252]