Validate instructions
Make sure the instruction is valid (as reported by the comgr
disassembler) before returning whether it is of a certain kind (branch,
cbranch, setpc, ...).

Validate the source and destination scalar operands of simulated
instructions.  Only a subset of registers is supported.

Add architecture_t::is_sequential.  Return false if the instruction is
using a non-sequential opcode even if the encoding is invalid.  This
is to avoid incorrectly reporting an instruction as sequential.
Invalid encodings should be reported as unknown.

Change-Id: Ic7f39d193741d7daf5b71ca1259242805efa5df4


[ROCm/rocdbgapi commit: 9129fb3548b0be9b412e03c1f2d912e20d679b9c]