Add support for gfx10's shared vgprs
The shared vgprs are 32-wide vector registers shared between the 2
halves of a wave64 on gfx10.  They are logically addressed right after
the 64-wide private vector registers.

Re-order the wave64 and wave32 registers so that the wave64 registers
are always reported first in wave64 mode with shared vgprs (registers
are reported in regnum ascending order).

Change-Id: I9698d32ba64fa750b197b67d71fe8d633ee2fd8b


[ROCm/rocdbgapi commit: d658fa611481fa6302d7d30f37f4101153b7b146]