Fix run-test.py
In the case of a memory exception, the KFD sends both a queue error and
and system error, and the ordering of those 2 events is not guaranteed.

Remove the 'System event' pattern matching as we could be dumping the
state of the waves as the result of a queue error.

Change-Id: I9aaa1a27e7c0a2275a9dc70529ab95a3567c34cf


[ROCm/rocr_debug_agent commit: a33f08a45e166b175ef9c833cd5c740d7f73a8a5]