Handle process exited exceptions
When a process exited exception is caught, stale queues, dispatches,
waves, and code objects are deleted and the API function is "retried"
to return a proper error code. For example, if a process exited
exception is thrown while reading a wave register, retrying the
function returns an invalid wave ID error.

Change-Id: I39bd4b50155a47a6f9693b3adc0f597ddb85c652


[ROCm/rocdbgapi commit: a3a25b81e11fe4119290929510181fe87dcb9411]