Handle deferred device memory violation exceptions
Device memory violations are either handled immediatly, if it can be
determined that no wavefront is the cause of the exception, or
deferred until a queue_wave_memory_violation exception is sent to the
runtime.

This change moves the clearing of the device memory violation exception
from wave_t::set_state (when the wave is resumed), to the process_t
when there are no more waves with memory violations.

Change-Id: I5a920b20aba2f340b593f100ba47bb44f61cba44


[ROCm/rocdbgapi commit: bd0f73d3b57ed33d33d825946605ad7d226a2c32]