doc: Rephrase parts about displaced single stepping
Fix and adjust some phrasings.

The terminology also sometimes feels strange.  Part of the documentation
is sometime about "single stepping buffers", and sometimes about "single
stepping".  As a consequence,  start/complete (from which the functions
are named) and allocate/release buffer (mostly used across the
documentation) are used to refer to the same thing. Sometimes there can
be a distinction, but having both terminologies is strange.  I feel
that the allocate/release makes more sense to me, but the API uses
start/complete, and I do not think it makes sense to update the API.

Change-Id: Iaa3b6b2b43fd6b5622781af9c4360c4f688b3859


[ROCm/rocdbgapi commit: 1b3d0c5e065ffcb6490613ab3fd41312b0d3eb13]