Fix dummy dispatch
When attaching to a process on an architecture that does not set up the
ttmp registers by default, an assertion is triggered when accessing the
dummy dispatch's kernel descriptor.

To avoid having to check that the dispatch is not the dummy dispatch
wherever the kernel descriptor is accessed, add a dummy kernel
descriptor class.

Change-Id: Ifa9748ccd9347dbb60d59a1450227a2a70fddc5a


[ROCm/rocdbgapi commit: 9452734d2d34fbb04d4779e5f50e7e6e22ecd68f]