Clear the PC's LSB when simulating instructions
Move the register masking to wave_t::write_register, it was previously
only enabled for API registers writes.  This allows the PC mask to be
applied any time the register is written, even if the wave is parked.

Because of the PC register read-only bits (lower bits are not wired and
therefore are always zero), a minimum instruction alignment is required.

Update the architecture's minimum_instruction_alignment to use the PC
register read-only mask to deduce the alignment requirement.

Change-Id: Ic1c37eba2823c46ad2f4b5b7e5bda29850b7bbfb


[ROCm/rocdbgapi commit: 49972860500619af4fc94805dd67c10e4633aa1f]