Add support for workgroups
Add a workgroup_t class to hold the coordinates of a group of waves.
All waves of a workgroup are associated with the same workgroup_t
instance.

All waves of a workgroup are guaranteed to be seen in a single context
save/restore, the hardware does not launch partial workgroups, so
workgroup_t can be marked and sweeped when decoding the control stack.

Remove the dispatch_grid and wave_in_group registers, and add new
cwsr_record_t queries to access them.

Add amd_dbgapi_workgroup_get_info and amd_dbgapi_workgroup_get_info.

Change-Id: Iccfe437300c5410cd6e8584c76fa4b2033209784


[ROCm/rocdbgapi commit: 21f78ce74fef83f55ef4c69f4bf1d7d1e42a5e7c]