Add to_cstring ()
to_cstring () is a convenience wrapper used to shorten
to_string ().c_str (). It constructs a temporary std::string in the
current expression's scope to hold the storage for a char array
corresponding to the content of to_string ().

Note that the returned char array pointer should not escape the scope
of the expression to_cstring is used in.  The same requirement was also
true for to_string ().c_str ().

Change-Id: I32b3039725cfda562aeed26a568f7d3f1433a07b


[ROCm/rocdbgapi commit: c00d96d51663d8ce82a92863d4a3cd93d9226f9e]