Fix freed memory access
When suspending the queue in a wave_t or workgroup_t member function,
it should be considered that the wave(s) could have exited and that the
current instance may have been destroyed by queue_t::update_queues.

The checks were in place, but the instance was still accessed after
queue_suspend to read the id.

Move reading the id before queue_suspend.  It is guaranteed to be read
before the instance is destroyed because of the fences on ::delete and
::free.

Change-Id: Ibd6caa0c2393ae57e4dedca43825f43633ba7d71


[ROCm/rocdbgapi commit: fc091c9e0ce227e4ada54eb68803ea2f2c36dbb4]