Add initialize_spi_ttmps and initialize_trap_handler_ttmps
initialize_spi_ttmps sets up the ttmp as SPI would if ttmps setup was
enabled.  initialize_trap_handler_ttmps sets up the ttmps as if the
trap handler had been called.

These 2 functions allow the wave_t::update implementation to abstract
architectures that depends on the trap handler being executed to be able
to read the wave's state saved in the ttmps.

Change-Id: I5e7f4e8997a9f7058859d1dfa21a282aafa42dd3


[ROCm/rocdbgapi commit: 8bf4d6540a38f4d3d2a8b60eb23aadf003973c02]