Fix queue_suspend/queue_resume
When a queue is all-stopped (it has no running waves, and it cannot
create new waves), the suspend and resume operations on that queue
are no-ops.

Since the order of queue_t`s is not preserved when building the array
of os_queue_id_t that is passed to the os_driver, the index of the
returned os_queue_id_t cannot be used to access the original queue_t.
Use std::find () instead.

Change-Id: I3ac2905cdcbbc7534a534c918b3fe04ed0a242fb


[ROCm/rocdbgapi commit: 82dda5be0cd68b2e284ab874a1640c0641259816]