Fix printing all wavefronts
When printing all wavefronts is requested, the debug agent must stop
running waves before inspecting their state.  The stop_all_wavefronts
function iterates all wavefronts, and stops the wavefronts that are
still running.

stop_all_wavefronts was failing to detect already stopped wavefronts.
Added a check before sending the stop request as it is illegal to
send a stop request to an already stopped wavefront.

Also noticed that the stopped events were not reported as processed,
and resuming wavefronts with stop events was returning an error.  A
wave is only trully stopped when the stop event is processed.  Added
amd_dbgapi_event_processed to both event loops.

Change-Id: I0eb09d61253e1682488eb6b3c389854f195fdab1


[ROCm/rocr_debug_agent commit: 0d7def50400b2a0bbc2ad34eb2b57d7f2ca3a745]