Remove faulty assertion when simulating 's_getpc_b64 vcc'
An assertion is triggered when simulating 's_getpc_b64 vcc' because
the amdgpu_regnum_t::vcc_lo register number is not naturally aligned.

The invariant really is that the encoding, not the regnum, for a
register pair should be aligned, as is_getpc () would not return true
if it wasn't.

Since this assertion is not strictly necessary (is_getpc () already
checks), we should remove it, but for completeness, also align the
regnum_t register pairs *_lo/hi on even numbers.

Change-Id: If5e102340c3ce940b4476db16b6052f10b6236e8


[ROCm/rocdbgapi commit: 23f80f8ba06ea82e89de540a90dce5ebdf530d85]