SWDEV-351980 - Fix hip_prof_str.h generation
Fixed an error in the CMakeLists.txt USE_PERF_API option declaration
that was making it always disabled. Fixing this exposed an issue with
the hip_prof_gen.py script's handling of function variants (for example,
_spt functions) and new HIP_INIT_API_* macros.

Also switched the python interpreter to python3 as python2 may not be
available by default on the build system.

Change-Id: I971fc9edcc746ca63a2bdf4f540e755f9a80fe69


[ROCm/clr commit: 8ac20e5e385856c97f2be7db82dc25e4bba151c6]