kfd_driver: Get agent name from pci_ids list
For future hardware generations, KFD will not report GPU name as they do today.
Instead "IP DISCOVERY" will be used.

This patch proposes to do a best effort to find a name based on the PCI vendor
and device IDs, based on the pci.ids database[1].  This is not entirely
foolproof and relies on the user to run `update-pciids` to fetch a hopefully
up-to-date database.

If the device cannot be found in the system database, a name is created
following the template "Device %04x:%04x".

The pci.ids file location is configured using CMake.  CMake will search the
file in well known locations used by major distros (/usr/share/misc/pci.dis and
/usr/share/hwdata/pci.ids).  For situations where the file is located somewhere
else on the system, the -DPCI_IDS_PATH=/path/to/pci.ids can be passed to CMake
to force the file location.

[1] https://pci-ids.ucw.cz/

Change-Id: I4ecfdf32202445cf3cfb61759589b68974e7d288


[ROCm/rocdbgapi commit: 197ea378c02ea2dac4f008c9ac6426bc08f7619a]