Don't resume the queues until after debug mode is disabled
There is a race condition between resuming the queues and disabling
debug mode. If the queues are resumed before debug mode is disabled,
then it is possible for a wave to raise an exception that will be
routed to the debugger instead of the runtime. Since the debugger is
in the process of detaching, that exception will be dropped.

The solution is to keep the queues suspended, then call the ioctl
to disable debug, the KFD will resume the queues after setting the
state to forward exceptions to the runtime.

Change-Id: Ic65d047780ba41352ab04d965b4e240b86eca156


[ROCm/rocdbgapi commit: 96b8bf3eebdbe05f4f5908c265f237770b7c07ab]