wave: remove the mutable m_parked_pc field
When dbgapi parks a wave, use the ttmp registers to save the original PC
instead of a member of wave_t.  This way, if the wave's state is saved
in a corefile (will be implemented in later patches), the original PC
can be recovered as if it has been saved by the trap handler.

This also remove the need of a mutable field in wave_t.

Change-Id: I25d020e59284c845bf158d8435ce8fe9fed49245


[ROCm/rocdbgapi commit: 21b3de245a593110ef9ba2ee1c10bf68523f0bba]