os_driver_t::send_exceptions: use std::optional agent and queue id
Future interface of KFD might drop the constants KFD_INVALID_GPUID and
KFD_INVALID_QUEUEID.  Those values are not really used in the KFD
interface (KFD never returns those), but are used internally in dbgapi.

The main use of those constants is to initialize the parameters for
os_driver_t::send_exceptions if the exception to send is not linked to
an agent or queue.  Instead of using the special values
os_invalid_agentid and os_invalid_queueid on the client side of the
call, use std::optional parameters.  The client can use std::nullopt
instead to carry the same information.

This moves the question of using "special" values away from process.cpp
to the implementation of kfd_driver_t::send_exception to initialize the
fields passed to KFD.  It is expected that if the exception sent is not
linked to an agent / a queue, the fields are ignored by KFD anyway.

Change-Id: Ia00dd75a166ed4a94133fc9005fc3d8c4dc77530


[ROCm/rocdbgapi commit: 321cbaff50db8f7bf1521102cc96ec657cc043c8]