queue_t::os_queue_id: do not return os_invalid_queueid
The queue_t::os_queue_id checks if the queue is valid before returning
the os_queue_id.  If the queue is invalid, it return os_invalid_queueid.

The os_invalid_queueid constant (KFD_INVALID_QUEUEID) might be removed
in future versions of KFD, so it might not be a well known value to to
use anymore.

Also, I find counter-intuitive to have non-unique ids (if multiple
queues are invalidated, then they share the same os_queue_id).

This patch proposes that queue_t::os_queue_id returns the original
queue_id with the invalid bit set (KFD_DBG_QUEUE_INVALID_BIT) if the
queue is invalid.

The main change is that two invalid queues will no longer have the same
os_queue_id, which is fine as this property was never used.

Change-Id: Id35e9d7bbd03710c0f780ea97317f9de0981bff4


[ROCm/rocdbgapi commit: 8dcc9ca49253720ec78179f26be5ea4c18392ba5]