os_queue_id: add the os_queue_id_unmask helper
Commit 5c59e0a (queue_t::os_queue_id: do not return os_invalid_queueid)
made queue_t::os_queue_id return the os_queue_id of the queue with the
invalid bit set for invalid queues.  When printing such id in logs, the
os_queue_id does not look like the "original" id.  Add the
os_queue_id_unmask helper function to remove unset invalid and error
bits (if set) in the os_queue_id and use it when appropriate.

Change-Id: Ic5728b224c6de565694a3f678577d32caf4a9025


[ROCm/rocdbgapi commit: 3d0c2d4c4e6c81706223fed5284237b8497cf5f2]