Use amd_dbgapi_notifier_t in os_driver_t::enable_debug
The os_driver_t::enable_debug method takes a parameter which is a "os
specific" handle the driver will be able to use to notify the debugger
that some event is available.  The current interface defined that this
parameter should be of type file_desc_t, but this is a linux
specificity.

This patch changes the parameter type to amd_dbgapi_notifier_t instead
so this type can represent what is used for notification on a given
target.  For Linux, both file_desc_t and amd_dbgapi_notifier_t are
aliases to int, so there is no ABI change with this patch.

Change-Id: If18ef7822044fb92bd9cdcd898e03b1bfe43dff1


[ROCm/rocdbgapi commit: 935cffd2ff2467806550b47fc19860ce6540ee96]