Do not use non portable ssize_t
The ssize_t type is not portable.  We have 2 avoidable uses of this
type.  In both instances, we cast the value returned by simm16_operand
(of type int16_t) to ssize_t before left-shifting it by 2.  The shift
operator automaticly implies an int promotion of the returned value to
int32_t.  However as we will add this to a uint64, we can change the
cast to int64_t directly.

Note that on Linux (64 bits), ssize_t to int64_t are equivalent, so this
patch is really a no-op.

While doing this modification, I also realized that some places did a
cast for similar reasons, but to std::ptrdiff_t.  As the casted offset
is really an offset in inferior memory and not in host memory, I also
convert those to use int64_t.

Note that similarly to our amd_dbgapi_global_address_t, we could add a
address difference type for inferior addresses:

    using amd_dbgapi_global_address_diff_t = int64_t;

and use this.  This might be done in a later patch.

Change-Id: I67cd3afcef91ec09082ca8676e64cdab0379a5bc


[ROCm/rocdbgapi commit: 0df02a4360e021456f487087455940ec02fdade9]