Add process_freeze / process_unfreeze methods
Add amd_dbgapi_process_freeze and amd_dbgapi_process_ununfreeze methods.

Those methods are used as part of corefile generation.  The freeze method
ensures that all activity is suspended on the agents and that any cached
data kept by dbgapi is flushed to memory.  This way after the state of the
system is captured by the client, another instance of dbgapi can reload
a consistent state.

Change-Id: Ia44b304ba02912554736f64c905366f1877ac7fe


[ROCm/rocdbgapi commit: a958e461a84c4a1a90f3a9359f53753823613432]