Ignore write_back memory if we do not have a real process
A following commit will add support to handle core dumps.  It is
possible that the underlying core dump is opened read-only, in which
case any write-back from the memory_cache_t will throw a memory_error.

This patch makes it so such error is ignored if the current process_t
does not map to a real OS process, i.e. its m_os_process_id does not
contain a value.

To ensure that after a call to write_back all cache lines are non-dirty,
this patch also makes it so memory_cache_t::write_back does not throw a
memory error on the fist write error.  Instead it tries to write all
dirty cache lines and throws one exception at the end of the procedure.

An alternative would be to remove the assertion in
memory_cache_t::discard which checks that no page is dirty at this
point.

Change-Id: Iaa8d5e7c72e967e34f1928f34ffa4ea816b11e72


[ROCm/rocdbgapi commit: 2dc5d8130f69f84e352fe6eb4b8ba9482440d7b8]