Park waves for gfx11
Starting rocr version 9, gfx11 parks waves.  Handle this in dbgapi.

Wave parking is required to handle situations where the wave is stopped
before a `s_sendmsg sendmsg(MSG_DEALLOC_VGPRS)` instruction.  In a
situation such a the following:

  s_trap 2
  s_sendmsg sendmsg(MSG_DEALLOC_VGPRS)
  s_endpgm

the shader enters trap handler at the s_trap instruction.  In the trap
handler, the wave's HALT bit is set and an interrupt is sent to the
driver to notify of the exception and returns to the shader.  At this
point the device prefetches the next instruction ("s_sendmsg"), and
because it recognizes it it considers VGPRS to be deallocated even if
the instruction has not executed yet.  In this state, when dbgapi stops
the wave, it requires a context save.  Because the hardware already
considers the VGPRS deallocated, it thinks that context save cannot be
done and kills the wave.

To prevent this "look at the next instruction" behavior, the trap
handler is modified to not return to the shader if the wave is halted.
Instead, it sets the wave's PC to point to an instruction where the wave
can remain halted and do a context-save.

Change-Id: I0b941038eed01b969200d37cc005b3f517df9e82


[ROCm/rocdbgapi commit: 57fbb9437dfeb664f7154ed6c3f6dac5771167f7]