test/run-test.py: Filter out precise memory warnings
Precise memory is enabled by default when running rocr_debug_agents.
However, all architectures do not support this feature, so a warning
might be printed to stderr.  When this happens, the test runner
considers this an error and reports the entire test as failed.

This patch ensures that a such warning printed to stderr does not cause
a test failure.

Tested on gfx900 and gfx1030.

Change-Id: Ife2d3966298521c51448e480596f5a934b311660


[ROCm/rocr_debug_agent commit: dfe793f5523fc4b881284cb3cf405329a879c2be]