Refactor workgroup detection/creation
When ttmps are not initialized, local memory is not accessible because
the workgroup the waves belong to is created with a zero sized local
storage.

The local storage size is available in the cwsr_record, so use it
when creating a new workgroup regardless of the availability of the
dispatch packet.

Change-Id: I9ed2beeb59ae50f1385793469674909ed90f334f


[ROCm/rocdbgapi commit: 2c1d2f19857370ef8f29c123e983c5007382389b]