Use the missed single-step workaround for gfx11
When mode.trap_after_inst_en is set, it is possible for a wave to enter
the trap handler, after single-stepping an instruction and a savectx is
raised, with only savectx set in trapsts.

Because trapsts.trap_after_inst is not reliably set, we still need to
use the missed single-step workaround for gfx11.

Change-Id: I95115d01a62efd2d5a897733d878cd441d63bdfc


[ROCm/rocdbgapi commit: 6d0814d9c2e91489255ce9f2dc5169937b44829c]